fixed blog router

+++ controllers/blog.js
@@ -2,7 +2,7 @@ const blogsRouter = require('express').Router();
 const Blogs = require('../models/blogs');
 const User = require('../models/user');
 
-const jwt = require('jsonwebtoken');
+const { userExtractor } = require('../utils/middleware');
 
 const getTokenFrom = (request) => {
   const authorization = request.get('authorization');
@@ -30,59 +30,48 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 });
 
-blogsRouter.post('/', async (request, response) => {
-  const body = request.body;
-
-  if (!body.title || !body.author || !body.url) {
-    return response.status(400).json({ error: 'Missing required fields' });
-  }
-
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' });
-  }
-
-  const user = await User.findById(decodedToken.id);
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body;
 
   const blog = new Blogs({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-    user: user.id,
+    title,
+    author,
+    url,
+    likes: likes ? likes : 0,
   });
 
-  const savedBlog = await blog.save();
-  user.blogs = user.blogs.concat(savedBlog._id);
+  const user = request.user;
+
+  if (!user) {
+    return response.status(401).json({ error: 'operation not permitted' });
+  }
+
+  blog.user = user._id;
+
+  const createdBlog = await blog.save();
+
+  user.blogs = user.blogs.concat(createdBlog._id);
   await user.save();
 
-  response.json(savedBlog);
+  response.status(201).json(createdBlog);
 });
 
-blogsRouter.delete('/:id', async (request, response) => {
-  try {
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
-    if (!decodedToken.id) {
-      return response.status(401).json({ error: 'Token invalid' });
-    }
-    
-    const blog = await Blogs.findById(request.params.id);
-    
-    if (!blog) {
-      return response.status(404).json({ error: 'Blog not found' });
-    }
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const blog = await Blogs.findById(request.params.id);
 
-    if (blog.user.toString() !== decodedToken.id) {
-      return response.status(403).json({ error: 'Unauthorized to delete this blog' });
-    }
+  const user = request.user;
 
-    await Blogs.findByIdAndRemove(request.params.id);
-    response.status(204).end();
-  } catch (error) {
-    response.status(500).json({ error: 'Server error' });
+  if (!user || blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'operation not permitted' });
   }
-});
 
+  user.blogs = user.blogs.filter((b) => b.toString() !== blog.id.toString());
+
+  await user.save();
+  await Blogs.findByIdAndRemove(request.params.id);
+
+  response.status(204).end();
+});
 
 blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body;
